backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
cloudinary==1.38.0
razorpay==1.4.1
firebase-admin==6.4.0
python-multipart==0.0.9
resend==0.7.0
python-dotenv==1.0.1
psycopg2-binary==2.9.11
slowapi==0.1.9
websockets==12.0
pydantic[email]==2.6.1
pydantic-settings==2.2.1
httpx==0.26.0
